crewai-test/requirements.txt
crewai>=0.95.0
python-dotenv>=1.0.1
pytest>=8.2.0
